# openapi.yaml
openapi: 3.0.3
info:
  title: WC-Hub API
  version: 1.0.0
  description: |
    WC-Hub é um control plane multi-cloud para gerenciamento unificado de infraestrutura.
    
    **Recursos principais:**
    - Gestão de VMs/containers (Proxmox, Docker, Kubernetes, OCI)
    - Infraestrutura como código (Terraform)
    - Integrations (GitHub, Cloudflare)
    - RBAC granular com policy enforcement
    - Audit logging completo
    - Terminal SSH via WebSocket
    
    **Autenticação:**
    - Session-based authentication com cookies HTTPOnly
    - TOTP 2FA opcional
    - Policy enforcement para ações críticas (X-Confirmation + X-TOTP-Code headers)
    
  contact:
    name: WC-Hub Development Team
  license:
    name: Proprietary
servers:
  - url: https://hub.example.com
    description: Production
  - url: http://localhost:8080
    description: Development
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/proxmox/nodes/{node}/{kind}/{vmid}/{action}:
    post:
      summary: Ação de energia em guest Proxmox (start, stop, shutdown, reboot)
      operationId: proxmoxPowerAction
      tags: [proxmox]
      parameters:
        - name: node
          in: path
          required: true
          schema:
            type: string
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum: [qemu, lxc]
        - name: vmid
          in: path
          required: true
          schema:
            type: integer
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [start, stop, shutdown, reboot, reset]
        - name: X-Confirmation
          in: header
          required: false
          schema:
            type: string
        - name: X-TOTP-Code
          in: header
          required: false
          schema:
            type: string
      responses:
        '202':
          description: Ação aceita
        '403':
          description: Ação bloqueada por policy
  /api/v1/proxmox/nodes/{node}/{kind}/{vmid}:
    delete:
      summary: Excluir VM ou LXC Proxmox
      operationId: proxmoxDeleteGuest
      tags: [proxmox]
      parameters:
        - name: node
          in: path
          required: true
          schema:
            type: string
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum: [qemu, lxc]
        - name: vmid
          in: path
          required: true
          schema:
            type: integer
        - name: X-Confirmation
          in: header
          required: true
          schema:
            type: string
        - name: X-TOTP-Code
          in: header
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Guest excluído
        '403':
          description: Ação bloqueada por policy
...
  /api/v1/jobs:
    get:
      summary: Listar jobs
      operationId: listJobs
      tags: [jobs]
      responses:
        '200':
          description: Lista de jobs
    post:
      summary: Criar/Enfileirar job
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '202':
          description: Job enfileirado
...
  /api/v1/terminal/tickets:
    post:
      summary: Criar ticket de sessão SSH Terminal
      operationId: createTerminalTicket
      tags: [terminal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_id, confirmation, totp_code]
              properties:
                host_id:
                  type: string
                confirmation:
                  type: string
                totp_code:
                  type: string
      responses:
        '201':
          description: Ticket gerado
        '403':
          description: Acesso bloqueado por policy
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...